+++ src/Headers/enemy.h
@@ -78,7 +78,7 @@ enum
 			
 ObjNode *MakeEnemySkeleton(Byte skeletonType, float x, float z, float scale, float rot, void (*moveCall)(ObjNode*));
 extern	void DeleteEnemy(ObjNode *theEnemy);
-Boolean DoEnemyCollisionDetect(ObjNode *theEnemy, unsigned long ctype, Boolean useBBoxBottom);
+Boolean DoEnemyCollisionDetect(ObjNode *theEnemy, uint32_t ctype, Boolean useBBoxBottom);
 extern	void UpdateEnemy(ObjNode *theNode);
 extern	void InitEnemyManager(void);
 void MoveEnemy(ObjNode *theNode);
