+++ src/Headers/sprites.h
@@ -13,9 +13,9 @@ enum
 
 typedef struct
 {
-	long				width,height;
+	int				width,height;
 	float			aspectRatio;			// h/w
-	long			srcFormat, destFormat;
+	int				srcFormat, destFormat;
 	MetaObjectPtr	materialObject;
 }SpriteType;
 
@@ -27,5 +27,5 @@ void LoadSpriteFile(FSSpec *spec, int groupNum);
 ObjNode *MakeSpriteObject(NewObjectDefinitionType *newObjDef);
 void BlendAllSpritesInGroup(short group);
 void ModifySpriteObjectFrame(ObjNode *theNode, short type);
-void DrawSprite(int	group, int type, float x, float y, float scale, float rot, u_long flags);
+void DrawSprite(int	group, int type, float x, float y, float scale, float rot, uint32_t flags);
 void BlendASprite(int group, int type);
